Separate routing into dealer and broker.
The routing has been moved as much as possible to the process handeling the client data.
So messages that do not need to be handled by routing code will be handled directly at
the client. This is e.g. the 'hello' message as it only needs to check wether a client is
allowed to connect and if it needs to authenticate itself.

Sending of events are also purely handled by the connection process, to reduce bottlenecks
as much as possible.

For calls and invocations a server is started to handle the data and in the future the
timeout etc.

The code to register/unregister and to subscribe and unsubscribe has been split into broker
and dealer, so the bottleneck there has been reduced to two processes instead of one.

The tests are passing and the coverage is pretty good, yet this is still work in progress
and needs some more implementation like the middleware and meta events and meta functions.