Set session when creating a session id.
When registering with sessions to gain a unique session id transmit
the realm the session will be connected to, so there won't be a
second message.
Also adding a function to send messages directly to sessions using
the ets-table of sessions, so all messages will be routed by ids
and not by pid. This will help when implementing router to router
communications.