Send messages from dealer and broker via erwa_sessions.
Use erwa_sessions to keep track of the session <-> pid mapping
and then use this mapping for sending messages.
Now dealer and broker only send messages to session ids,
not to pids. The result is the possibility to easily route messages
eg through a bridge to another router or similar.

The only exception is erwa_routing, whose two messages are still
sent directly to the connected pids, yet they will stay local, so
there is no need to use the mapping.